+++ jlib-persistence-hibernate/src/main/java/org/jlib/persistence/hibernate/usertype/primitive/IdEnumCharUserType.java
@@ -38,7 +38,7 @@ import org.jlib.persistence.hibernate.usertype.InvalidUserTypeParametersExceptio
 import static org.jlib.persistence.hibernate.usertype.UserTypeUtility.assertValidParametersCount;
 import static org.jlib.persistence.hibernate.usertype.UserTypeUtility.getParameterValue;
 import org.jlib.persistence.IdEnum;
-import org.jlib.reflect.programelement.ProgramElementException;
+import org.jlib.reflect.languageelement.ProgramElementException;
 import static org.jlib.reflect.reflector.defaults.Reflectors.useClass;
 
 public class IdEnumCharUserType<EnumValue extends Enum<EnumValue> & IdEnum<Id>, Id extends Serializable>
