+++ jlib-reflect/src/main/java/org/jlib/reflect/languageelement/ClassInstantiationException.java
@@ -19,7 +19,7 @@
  *     limitations under the License.
  */
 
-package org.jlib.reflect.programelement;
+package org.jlib.reflect.languageelement;
 
 import org.jlib.message.Message;
