+++ jlib-reflect/src/main/java/org/jlib/reflect/languageelement/InvalidMethodParameterTypesException.java
@@ -19,18 +19,22 @@
  *     limitations under the License.
  */
 
-package org.jlib.reflect.programelement;
-
-import java.lang.reflect.Method;
+package org.jlib.reflect.languageelement;
 
 import static org.jlib.message.MessageUtility.message;
 
-public class ConcreteMethodExpectedException
+public class InvalidMethodParameterTypesException
 extends MethodLookupException {
 
-    private static final long serialVersionUID = - 5410080091550763771L;
+    private static final long serialVersionUID = - 4942289385749320071L;
 
-    public ConcreteMethodExpectedException(final Method method) {
-        super(message(), method);
+    public InvalidMethodParameterTypesException(final String className, final String methodName,
+                                                final Class<?>[] parameterTypes) {
+        super(message(), className, methodName, parameterTypes);
+    }
+
+    public InvalidMethodParameterTypesException(final String className, final String methodName,
+                                                final Class<?>[] parameterTypes, final Exception cause) {
+        super(message(), className, methodName, parameterTypes, cause);
     }
 }
