+++ jlib-reflect/src/main/java/org/jlib/reflect/languageelement/LanguageElementUtility.java
@@ -19,7 +19,7 @@
  *     limitations under the License.
  */
 
-package org.jlib.reflect.programelement;
+package org.jlib.reflect.languageelement;
 
 import java.lang.reflect.Executable;
