+++ jlib-reflect/src/main/java/org/jlib/reflect/reflector/TypedMethod2.java
@@ -23,7 +23,7 @@ package org.jlib.reflect.reflector;
 
 import java.lang.reflect.Executable;
 
-import org.jlib.reflect.programelement.MethodLookupException;
+import org.jlib.reflect.languageelement.MethodLookupException;
 
 public interface TypedMethod2<Exe extends Executable, ReturnType, Argument1, Argument2>
     extends TypedMethod<Exe, ReturnType> {
