+++ jlib-reflect/src/main/java/org/jlib/reflect/reflector/TypedMethod3.java
@@ -23,7 +23,7 @@ package org.jlib.reflect.reflector;
 
 import java.lang.reflect.Executable;
 
-import org.jlib.reflect.programelement.MethodLookupException;
+import org.jlib.reflect.languageelement.MethodLookupException;
 
 public interface TypedMethod3<Exe extends Executable, ReturnType, Argument1, Argument2, Argument3>
     extends TypedMethod<Exe, ReturnType> {
