+++ jlib-reflect/src/main/java/org/jlib/reflect/reflector/defaults/invoke/ConstructorInvokeStrategy.java
@@ -23,8 +23,8 @@ package org.jlib.reflect.reflector.defaults.invoke;
 
 import java.lang.reflect.Constructor;
 
-import org.jlib.reflect.programelement.LanguageElementHelper;
-import org.jlib.reflect.programelement.MethodInvocationException;
+import org.jlib.reflect.languageelement.LanguageElementHelper;
+import org.jlib.reflect.languageelement.MethodInvocationException;
 
 public class ConstructorInvokeStrategy<InstantiatedObject>
     extends InvokeStrategy<Constructor<InstantiatedObject>> {
