+++ jlib-reflect/src/main/java/org/jlib/reflect/reflector/defaults/methodreturn/DefaultMethodReturn.java
@@ -27,10 +27,10 @@ import static java.util.Arrays.asList;
 import static lombok.AccessLevel.PROTECTED;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.jlib.reflect.programelement.InvalidReturnTypeException;
-import org.jlib.reflect.programelement.LanguageElementHelper;
-import static org.jlib.reflect.programelement.LanguageElementUtility.assertReturnTypeInstanceOf;
-import org.jlib.reflect.programelement.ProgramElementException;
+import org.jlib.reflect.languageelement.InvalidReturnTypeException;
+import org.jlib.reflect.languageelement.LanguageElementHelper;
+import static org.jlib.reflect.languageelement.LanguageElementUtility.assertReturnTypeInstanceOf;
+import org.jlib.reflect.languageelement.ProgramElementException;
 import org.jlib.reflect.reflector.MethodOverload;
 import org.jlib.reflect.reflector.MethodReturn;
 import org.jlib.reflect.reflector.defaults.overload.DefaultInstanceMethodOverload;
